modules/group-iv-adaptive-systems/metacognition/pyproject.toml
[project]
name = "metacognition"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-exporter-prometheus>=0.45b0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "prometheus-client>=0.20.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "structlog>=24.1",
    "httpx>=0.27",
    "endogenai-vector-store",
    "endogenai-a2a>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "mypy>=1.10.0", "ruff>=0.4.0"]

[tool.uv.sources]
endogenai-vector-store = { path = "../../../shared/vector-store/python", editable = true }
endogenai-a2a = { path = "../../../shared/a2a/python", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metacognition"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: marks tests requiring live services",
]

[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
